+++ test/charset_test.dart
@@ -14,6 +14,42 @@ void main() {
         utf16.encode("上善若水"), [254, 255, 78, 10, 85, 132, 130, 229, 108, 52]);
   });
 
+  test('detect', () {
+    String text = "上善若水";
+
+    List<int> gbkData = gbk.encode(text);
+    expect(gbk.name, Charset.detect(gbkData)?.name);
+
+    List<int> eucJpData = eucJp.encode(text);
+    expect(
+        eucJp.name,
+        Charset.detect(
+          eucJpData,
+          orders: [ascii, eucJp, latin1],
+        )?.name);
+
+    List<int> shiftJisData = shiftJis.encode(text);
+    expect(
+        shiftJis.name,
+        Charset.detect(
+          shiftJisData,
+          orders: [ascii, shiftJis, latin1],
+        )?.name);
+
+    text = '상선이 물과 같다';
+    List<int> eucKrData = eucKr.encode(text);
+    expect(
+        eucKr.name,
+        Charset.detect(
+          eucKrData,
+          orders: [ascii, eucKr, latin1],
+        )?.name);
+
+    text = 'ABC55/';
+    List<int> asciiData = ascii.encode(text);
+    expect(ascii.name, Charset.detect(asciiData)?.name);
+  });
+
   test('controls', () {
     String textl = "0x0atdaa\t\rRsad";
     expect(latin1.encode(textl), latin2.encode(textl));